feat : separate usecase
