+++ src/main/kotlin/kr/hhplus/be/server/application/port/in/concert/GetConcertListUseCase.kt
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.application.port.`in`.concert
+
+import kr.hhplus.be.server.application.dto.concert.ConcertResult
+
+interface GetConcertListUseCase {
+    fun getConcertList(): List<ConcertResult>
+}
