+++ src/main/kotlin/kr/hhplus/be/server/application/port/in/payment/ProcessPaymentUseCase.kt
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.application.port.`in`.payment
+
+import kr.hhplus.be.server.application.dto.payment.PaymentResult
+import kr.hhplus.be.server.application.dto.payment.ProcessPaymentCommand
+
+interface ProcessPaymentUseCase {
+    fun processPayment(command: ProcessPaymentCommand): PaymentResult
+}
