+++ src/main/kotlin/kr/hhplus/be/server/application/port/in/queue/CompleteTokenUseCase.kt
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.application.port.`in`.queue
+
+import kr.hhplus.be.server.application.dto.queue.CompleteTokenCommand
+
+interface CompleteTokenUseCase {
+    fun completeToken(command: CompleteTokenCommand): Boolean
+}
