+++ src/main/kotlin/kr/hhplus/be/server/application/port/in/user/ChargeUserPointUseCase.kt
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.application.port.`in`.user
+
+import kr.hhplus.be.server.application.dto.user.ChargeUserPointCommand
+import kr.hhplus.be.server.application.dto.user.UserResult
+
+interface ChargeUserPointUseCase {
+    fun chargeUserPoint(command: ChargeUserPointCommand): UserResult
+}
