+++ src/main/kotlin/kr/hhplus/be/server/application/port/in/user/UseUserPointUseCase.kt
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.application.port.`in`.user
+
+import kr.hhplus.be.server.application.dto.user.UseUserPointCommand
+import kr.hhplus.be.server.application.dto.user.UserResult
+
+interface UseUserPointUseCase {
+    fun useUserPoint(command: UseUserPointCommand): UserResult
+}
