+++ src/main/kotlin/kr/hhplus/be/server/application/service/user/UserQueryService.kt
@@ -3,7 +3,7 @@ package kr.hhplus.be.server.application.service.user
 import kr.hhplus.be.server.application.dto.user.GetUserCommand
 import kr.hhplus.be.server.application.dto.user.UserResult
 import kr.hhplus.be.server.application.mapper.UserMapper
-import kr.hhplus.be.server.application.port.`in`.GetUserUseCase
+import kr.hhplus.be.server.application.port.`in`.user.GetUserUseCase
 import kr.hhplus.be.server.application.port.out.queue.UserRepository
 import kr.hhplus.be.server.domain.users.UserDomainService
 import org.springframework.stereotype.Service
