+++ src/test/java/kr/hhplus/be/server/service/PaymentServiceTest.kt
@@ -8,11 +8,11 @@ import kr.hhplus.be.server.application.dto.payment.GetUserPaymentsCommand
 import kr.hhplus.be.server.application.dto.payment.ProcessPaymentCommand
 import kr.hhplus.be.server.application.dto.queue.ValidateTokenCommand
 import kr.hhplus.be.server.application.dto.queue.ValidateTokenResult
-import kr.hhplus.be.server.application.port.`in`.CompleteTokenUseCase
-import kr.hhplus.be.server.application.port.`in`.GetPaymentUseCase
-import kr.hhplus.be.server.application.port.`in`.GetUserPaymentsUseCase
-import kr.hhplus.be.server.application.port.`in`.ProcessPaymentUseCase
-import kr.hhplus.be.server.application.port.`in`.ValidateTokenUseCase
+import kr.hhplus.be.server.application.port.`in`.queue.CompleteTokenUseCase
+import kr.hhplus.be.server.application.port.`in`.payment.GetPaymentUseCase
+import kr.hhplus.be.server.application.port.`in`.payment.GetUserPaymentsUseCase
+import kr.hhplus.be.server.application.port.`in`.payment.ProcessPaymentUseCase
+import kr.hhplus.be.server.application.port.`in`.queue.ValidateTokenUseCase
 import kr.hhplus.be.server.application.port.out.concert.ConcertSeatGradeRepository
 import kr.hhplus.be.server.application.port.out.concert.ConcertSeatRepository
 import kr.hhplus.be.server.application.port.out.log.PointHistoryRepository
@@ -29,7 +29,6 @@ import kr.hhplus.be.server.domain.payment.Payment
 import kr.hhplus.be.server.domain.payment.exception.PaymentAlreadyProcessedException
 import kr.hhplus.be.server.domain.payment.exception.PaymentNotFoundException
 import kr.hhplus.be.server.domain.reservation.TempReservation
-import kr.hhplus.be.server.domain.reservation.TempReservationStatus
 import kr.hhplus.be.server.domain.reservation.exception.TempReservationNotFoundException
 import kr.hhplus.be.server.domain.users.User
 import kr.hhplus.be.server.domain.users.exception.InsufficientPointException
