+++ src/test/java/kr/hhplus/be/server/service/ReservationCommandServiceTest.kt
@@ -8,12 +8,12 @@ import kr.hhplus.be.server.application.dto.queue.ValidateTokenResult
 import kr.hhplus.be.server.application.dto.reservation.CancelReservationCommand
 import kr.hhplus.be.server.application.dto.reservation.ConfirmTempReservationCommand
 import kr.hhplus.be.server.application.dto.reservation.TempReservationCommand
-import kr.hhplus.be.server.application.port.`in`.CancelReservationUseCase
-import kr.hhplus.be.server.application.port.`in`.CompleteTokenUseCase
-import kr.hhplus.be.server.application.port.`in`.ConfirmTempReservationUseCase
-import kr.hhplus.be.server.application.port.`in`.ExpireTokenUseCase
-import kr.hhplus.be.server.application.port.`in`.TempReservationUseCase
-import kr.hhplus.be.server.application.port.`in`.ValidateTokenUseCase
+import kr.hhplus.be.server.application.port.`in`.reservation.CancelReservationUseCase
+import kr.hhplus.be.server.application.port.`in`.queue.CompleteTokenUseCase
+import kr.hhplus.be.server.application.port.`in`.reservation.ConfirmTempReservationUseCase
+import kr.hhplus.be.server.application.port.`in`.queue.ExpireTokenUseCase
+import kr.hhplus.be.server.application.port.`in`.reservation.TempReservationUseCase
+import kr.hhplus.be.server.application.port.`in`.queue.ValidateTokenUseCase
 import kr.hhplus.be.server.application.port.out.concert.ConcertSeatRepository
 import kr.hhplus.be.server.application.port.out.queue.UserRepository
 import kr.hhplus.be.server.application.port.out.reservation.ReservationRepository
