+++ src/test/java/kr/hhplus/be/server/service/UserServiceTest.kt
@@ -6,9 +6,9 @@ import io.mockk.verify
 import kr.hhplus.be.server.application.dto.user.ChargeUserPointCommand
 import kr.hhplus.be.server.application.dto.user.GetUserCommand
 import kr.hhplus.be.server.application.dto.user.UseUserPointCommand
-import kr.hhplus.be.server.application.port.`in`.ChargeUserPointUseCase
-import kr.hhplus.be.server.application.port.`in`.GetUserUseCase
-import kr.hhplus.be.server.application.port.`in`.UseUserPointUseCase
+import kr.hhplus.be.server.application.port.`in`.user.ChargeUserPointUseCase
+import kr.hhplus.be.server.application.port.`in`.user.GetUserUseCase
+import kr.hhplus.be.server.application.port.`in`.user.UseUserPointUseCase
 import kr.hhplus.be.server.application.port.out.queue.UserRepository
 import kr.hhplus.be.server.application.service.user.UserCommandService
 import kr.hhplus.be.server.application.service.user.UserQueryService
